backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
firebase-admin==6.5.0
face-recognition==1.3.0
opencv-python==4.10.0.84
numpy==2.1.2
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
dlib==19.24.6
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2 